Release docker-zulip 2.0.4-1.
This release is being done because 2.0.4-0 was incorrectly build off
the wrong commit in zulip/zulip (due to a mispushed tag), and has
about 1000 commits of development from master in it.

Pushing this will fix the situation for folks who are doing new
installations; we recommend folks running 2.0.4-0 not upgrade.

Fixes #201.