entrypoint: Rework certificate handling.
In Zulip Server 4.9-0, we started unconditionally installing
`certbot`, leading to `/etc/letsencrypt` always existing in the
container, and breaking the symlinking of it into `/data/`, and
certbot certificates in general.

Rework how certificates are obtained and stored.  Certificates are now
stored in subdirectories of `/data/certs/`, so that auto-generated
certificates (either from certbot, or self-signed) cannot ever
overwrite manual certificates.  Only the data parts of
`/etc/letsencrypt` are stored in the data volume; the general
configuration aspects and hooks are left for Puppet to configure.

We swap `certbot-deploy-hook` for overwriting the
`/etc/letsencrypt/renewal-hooks/deploy/050-nginx.sh` hook, leaving the
other hooks (for symlinking, and restarting the email server)
untouched.  The core Zulip Server behaviour of symlinking the certbot
certificates into `/etc/ssl/` is left untouched; we deal with
persistence and backup by dint of storing the `/etc/letsencrypt` in
the data volume.

Fixes #381.
Fixes #489.